Move ranking/score calcs from service files to job files
