+++ server/src/api/services/internal/group.service.ts
@@ -10,7 +10,6 @@ import { get200msCount, getCombatLevel, getLevel, getTotalLevel } from '../../ut
 import { getMeasure, getRankKey, getValueKey, isSkill } from '../../util/metrics';
 import * as cryptService from '../external/crypt.service';
 import * as achievementService from './achievement.service';
-import * as competitionService from './competition.service';
 import * as deltaService from './delta.service';
 import * as playerService from './player.service';
 import * as recordService from './record.service';
@@ -787,92 +786,6 @@ async function getOutdatedMembers(groupId) {
   return membersToUpdate.map(({ player }) => player);
 }
 
-async function refreshScores() {
-  const allGroups = await Group.findAll();
-
-  await Promise.all(
-    allGroups.map(async group => {
-      const currentScore = group.score;
-      const newScore = await calculateScore(group);
-
-      if (newScore !== currentScore) {
-        await group.update({ score: newScore });
-      }
-    })
-  );
-}
-
-async function calculateScore(group: Group): Promise<number> {
-  let score = 0;
-
-  const now = new Date();
-  const members = await getMembersList(group);
-  const pagination = { limit: 100, offset: 0 };
-  const competitions = await competitionService.getGroupCompetitions(group.id, pagination);
-  const averageOverallExp = members.reduce((acc: any, cur: any) => acc + cur, 0) / members.length;
-
-  if (!members || members.length === 0) {
-    return score;
-  }
-
-  // If has atleast one leader
-  if (members.filter(m => m.role === 'leader').length >= 1) {
-    score += 30;
-  }
-
-  // If has atleast 10 players
-  if (members.length >= 10) {
-    score += 20;
-  }
-
-  // If has atleast 50 players
-  if (members.length >= 50) {
-    score += 40;
-  }
-
-  // If average member overall exp > 30m
-  if (averageOverallExp >= 30000000) {
-    score += 30;
-  }
-
-  // If average member overall exp > 100m
-  if (averageOverallExp >= 100000000) {
-    score += 60;
-  }
-
-  // If has a clan chat
-  if (group.clanChat && group.clanChat.length > 0) {
-    score += 50;
-  }
-
-  // If has a description
-  if (group.description && group.description.length > 0) {
-    score += 40;
-  }
-
-  // If has a homeworld
-  if (group.homeworld && group.homeworld > 0) {
-    score += 20;
-  }
-
-  // If is verified (clan leader is in our discord)
-  if (group.verified) {
-    score += 100;
-  }
-
-  // If has atleast one ongoing competition
-  if (competitions.filter(c => c.startsAt <= now && c.endsAt >= now).length >= 1) {
-    score += 50;
-  }
-
-  // If has atleast one upcoming competition
-  if (competitions.filter(c => c.startsAt >= now).length >= 1) {
-    score += 30;
-  }
-
-  return score;
-}
-
 export {
   resolve,
   getMembers,
@@ -893,6 +806,5 @@ export {
   addMembers,
   removeMembers,
   changeRole,
-  updateAllMembers,
-  refreshScores
+  updateAllMembers
 };
